+++ UltitemsCyan/UltAssets.cs
@@ -40,19 +40,21 @@ namespace UltitemsCyan
         public static Sprite CorrodingVaultSprite;
         //public static Sprite FaultyBulbSprite;
         public static Sprite GrapevineSprite;
+        public static Sprite PigsSporkSprite;
         public static Sprite RockyTaffySprite;
         public static Sprite SuesMandiblesSprite;
         public static Sprite ViralSmogSprite;
         public static GameObject CorrodingVaultPrefab;
         //public static GameObject FaultyBulbPrefab;
         public static GameObject GrapevinePrefab;
+        public static GameObject PigsSporkPrefab;
         public static GameObject RockyTaffyPrefab;
         public static GameObject SuesMandiblesPrefab;
         public static GameObject ViralSmogPrefab;
 
 
         // Void
-        public static Sprite ChrysotopeSprite;
+        public static Sprite CrysotopeSprite;
         public static Sprite DownloadedRAMSprite;
         public static Sprite DriedHamSprite;
         public static Sprite InhabitedCoffinSprite;
@@ -62,7 +64,7 @@ namespace UltitemsCyan
         //public static Sprite WormHolesSprite;
         public static Sprite ZorsePillSprite;
         public static GameObject DownloadedRAMPrefab;
-        public static GameObject ChrysotopePrefab;
+        public static GameObject CrysotopePrefab;
         public static GameObject DriedHamPrefab;
         public static GameObject InhabitedCoffinPrefab;
         public static GameObject JubilantFoePrefab;
@@ -121,13 +123,14 @@ namespace UltitemsCyan
 
         // Buffs
         public static Sprite BirthdaySprite;
-        public static Sprite ChrysotopeFlySprite;
+        public static Sprite CrysotopeFlySprite;
         public static Sprite DownloadedSprite;
         public static Sprite DreamSpeedSprite;
         public static Sprite FrisbeeGlideSprite;
         public static Sprite GrapeSprite;
         public static Sprite OverclockedSprite;
         public static Sprite RottingSprite;
+        public static Sprite SporkBleedSprite;
         public static Sprite TickCritSprite;
         public static Sprite ZorseStarveSprite;
 
@@ -140,12 +143,12 @@ namespace UltitemsCyan
             // For Local Testing
             if (mainBundle == null)
             {
-                Log.Warning("Null Bundle");
+                Log.Warning("Null Bundle... getting for Debug");
                 path = System.IO.Path.Combine(System.IO.Path.GetDirectoryName(System.IO.Path.GetDirectoryName(Ultitems.PInfo.Location)), "assetbundle", "ultitembundle");
                 mainBundle = AssetBundle.LoadFromFile(path);
             }
 
-            float localScale = 0.1f;
+            //float localScale = 0.1f;
 
             // * * * White * * * 
             CremeBruleeSprite = mainBundle.LoadAsset<Sprite>("CremeBrulee.png");
@@ -156,14 +159,15 @@ namespace UltitemsCyan
             CremeBruleePrefab = mainBundle.LoadAsset<GameObject>("CremeBrulee.prefab");
             FleaBagPrefab = mainBundle.LoadAsset<GameObject>("FleaBag.prefab");
             FrisbeePrefab = mainBundle.LoadAsset<GameObject>("Frisbee.prefab");
+            //FrisbeePrefab = mainBundle.LoadAsset<GameObject>("Frisbee3D.prefab");
             KoalaStickerPrefab = mainBundle.LoadAsset<GameObject>("KoalaSticker.prefab");
             ToyRobotPrefab = mainBundle.LoadAsset<GameObject>("ToyRobot.prefab");
 
-            CremeBruleePrefab.transform.localScale = Vector3.up * localScale;
-            FleaBagPrefab.transform.localScale = Vector3.up * localScale;
-            FrisbeePrefab.transform.localScale = Vector3.up * localScale;
-            KoalaStickerPrefab.transform.localScale = Vector3.up * localScale;
-            ToyRobotPrefab.transform.localScale = Vector3.up * localScale;
+            //CremeBruleePrefab.transform.localScale = Vector3.up * localScale;
+            //FleaBagPrefab.transform.localScale = Vector3.up * localScale;
+            //FrisbeePrefab.transform.localScale = Vector3.up * localScale;
+            //KoalaStickerPrefab.transform.localScale = Vector3.up * localScale;
+            //ToyRobotPrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Green * * * 
             BirthdayCandleSprite = mainBundle.LoadAsset<Sprite>("BirthdayCandles.png");
@@ -179,36 +183,38 @@ namespace UltitemsCyan
             //TippedArrowPrefab = mainBundle.LoadAsset<GameObject>("TippedArrow.prefab");
             XenonAmpoulePrefab = mainBundle.LoadAsset<GameObject>("XenonAmpoule.prefab");
 
-            BirthdayCandlePrefab.transform.localScale = Vector3.up * localScale;
-            DegreeScissorsPrefab.transform.localScale = Vector3.up * localScale;
-            HMTPrefab.transform.localScale = Vector3.up * localScale;
-            OverclockedGPUPrefab.transform.localScale = Vector3.up * localScale;
+            //BirthdayCandlePrefab.transform.localScale = Vector3.up * localScale;
+            //DegreeScissorsPrefab.transform.localScale = Vector3.up * localScale;
+            //HMTPrefab.transform.localScale = Vector3.up * localScale;
+            //OverclockedGPUPrefab.transform.localScale = Vector3.up * localScale;
             //TippedArrowPrefab.transform.localScale = Vector3.up * localScale;
-            XenonAmpoulePrefab.transform.localScale = Vector3.up * localScale;
+            //XenonAmpoulePrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Red * * * 
             CorrodingVaultSprite = mainBundle.LoadAsset<Sprite>("CorrodingVault.png");
             //FaultyBulbSprite = mainBundle.LoadAsset<Sprite>("FaultyBulb.png");
             GrapevineSprite = mainBundle.LoadAsset<Sprite>("Grapevine.png");
+            PigsSporkSprite = mainBundle.LoadAsset<Sprite>("PigsSpork.png");
             RockyTaffySprite = mainBundle.LoadAsset<Sprite>("RockyTaffy.png");
             SuesMandiblesSprite = mainBundle.LoadAsset<Sprite>("SuesMandibles.png");
             ViralSmogSprite = mainBundle.LoadAsset<Sprite>("ViralSmog.png");
             CorrodingVaultPrefab = mainBundle.LoadAsset<GameObject>("CorrodingVault.prefab");
             //FaultyBulbPrefab = mainBundle.LoadAsset<GameObject>("FaultyBulb.prefab");
             GrapevinePrefab = mainBundle.LoadAsset<GameObject>("Grapevine.prefab");
+            PigsSporkPrefab = mainBundle.LoadAsset<GameObject>("PigsSpork.prefab");
             RockyTaffyPrefab = mainBundle.LoadAsset<GameObject>("RockyTaffy.prefab");
             SuesMandiblesPrefab = mainBundle.LoadAsset<GameObject>("SuesMandibles.prefab");
             ViralSmogPrefab = mainBundle.LoadAsset<GameObject>("ViralSmog.prefab");
 
-            CorrodingVaultPrefab.transform.localScale = Vector3.up * localScale;
+            //CorrodingVaultPrefab.transform.localScale = Vector3.up * localScale;
             //FaultyBulbPrefab.transform.localScale = Vector3.up * localScale;
-            GrapevinePrefab.transform.localScale = Vector3.up * localScale;
-            RockyTaffyPrefab.transform.localScale = Vector3.up * localScale;
-            SuesMandiblesPrefab.transform.localScale = Vector3.up * localScale;
-            ViralSmogPrefab.transform.localScale = Vector3.up * localScale;
+            //GrapevinePrefab.transform.localScale = Vector3.up * localScale;
+            //RockyTaffyPrefab.transform.localScale = Vector3.up * localScale;
+            //SuesMandiblesPrefab.transform.localScale = Vector3.up * localScale;
+            //ViralSmogPrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Void * * * 
-            ChrysotopeSprite = mainBundle.LoadAsset<Sprite>("Chrysotope.png");
+            CrysotopeSprite = mainBundle.LoadAsset<Sprite>("Chrysotope.png");
             DownloadedRAMSprite = mainBundle.LoadAsset<Sprite>("DownloadedRAM.png");
             DriedHamSprite = mainBundle.LoadAsset<Sprite>("DriedHam.png");
             InhabitedCoffinSprite = mainBundle.LoadAsset<Sprite>("InhabitedCoffin.png");
@@ -217,7 +223,7 @@ namespace UltitemsCyan
             //TungstenRodSprite = mainBundle.LoadAsset<Sprite>("TungstenRod.png");
             //WormHolesSprite = mainBundle.LoadAsset<Sprite>("WormHoles.png");
             ZorsePillSprite = mainBundle.LoadAsset<Sprite>("ZorsePill.png");
-            ChrysotopePrefab = mainBundle.LoadAsset<GameObject>("Chrysotope.prefab");
+            CrysotopePrefab = mainBundle.LoadAsset<GameObject>("Chrysotope.prefab");
             DownloadedRAMPrefab = mainBundle.LoadAsset<GameObject>("DownloadedRAM.prefab");
             DriedHamPrefab = mainBundle.LoadAsset<GameObject>("DriedHam.prefab");
             InhabitedCoffinPrefab = mainBundle.LoadAsset<GameObject>("InhabitedCoffin.prefab");
@@ -227,15 +233,15 @@ namespace UltitemsCyan
             //WormHolesPrefab = mainBundle.LoadAsset<GameObject>("WormHoles.prefab");
             ZorsePillPrefab = mainBundle.LoadAsset<GameObject>("ZorsePill.prefab");
 
-            ChrysotopePrefab.transform.localScale = Vector3.up * localScale;
-            DownloadedRAMPrefab.transform.localScale = Vector3.up * localScale;
-            DriedHamPrefab.transform.localScale = Vector3.up * localScale;
-            InhabitedCoffinPrefab.transform.localScale = Vector3.up * localScale;
-            JubilantFoePrefab.transform.localScale = Vector3.up * localScale;
-            RottenBonesPrefab.transform.localScale = Vector3.up * localScale;
+            //CrysotopePrefab.transform.localScale = Vector3.up * localScale;
+            //DownloadedRAMPrefab.transform.localScale = Vector3.up * localScale;
+            //DriedHamPrefab.transform.localScale = Vector3.up * localScale;
+            //InhabitedCoffinPrefab.transform.localScale = Vector3.up * localScale;
+            //JubilantFoePrefab.transform.localScale = Vector3.up * localScale;
+            //RottenBonesPrefab.transform.localScale = Vector3.up * localScale;
             //TungstenRodPrefab.transform.localScale = Vector3.up * localScale;
             //WormHolesPrefab.transform.localScale = Vector3.up * localScale;
-            ZorsePillPrefab.transform.localScale = Vector3.up * localScale;
+            //ZorsePillPrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Lunar * * * 
             //CreatureDeckSprite = mainBundle.LoadAsset<Sprite>("CreatureDeck.png");
@@ -252,11 +258,11 @@ namespace UltitemsCyan
             SilverThreadPrefab = mainBundle.LoadAsset<GameObject>("SilverThread.prefab");
 
             //CreatureDeckPrefab.transform.localScale = Vector3.up * localScale;
-            DreamFuelPrefab.transform.localScale = Vector3.up * localScale;
-            UltravioletBulbPrefab.transform.localScale = Vector3.up * localScale;
+            //DreamFuelPrefab.transform.localScale = Vector3.up * localScale;
+            //UltravioletBulbPrefab.transform.localScale = Vector3.up * localScale;
             //PowerChipsPrefab.transform.localScale = Vector3.up * localScale;
-            SandPailPrefab.transform.localScale = Vector3.up * localScale;
-            SilverThreadPrefab.transform.localScale = Vector3.up * localScale;
+            //SandPailPrefab.transform.localScale = Vector3.up * localScale;
+            //SilverThreadPrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Untiered * * * 
             CorrodingVaultConsumedSprite = mainBundle.LoadAsset<Sprite>("CorrodingVaultConsumed.png");
@@ -270,11 +276,11 @@ namespace UltitemsCyan
             SuesMandiblesConsumedPrefab = mainBundle.LoadAsset<GameObject>("SuesMandiblesConsumed.prefab");
             UniversalSolventPrefab = mainBundle.LoadAsset<GameObject>("UniversalSolvent.prefab");
 
-            CorrodingVaultConsumedPrefab.transform.localScale = Vector3.up * localScale;
-            InhabitedCoffinConsumedPrefab.transform.localScale = Vector3.up * localScale;
-            SuesMandiblesConsumedPrefab.transform.localScale = Vector3.up * localScale;
-            SilverThreadConsumedPrefab.transform.localScale = Vector3.up * localScale;
-            UniversalSolventPrefab.transform.localScale = Vector3.up * localScale;
+            //CorrodingVaultConsumedPrefab.transform.localScale = Vector3.up * localScale;
+            //InhabitedCoffinConsumedPrefab.transform.localScale = Vector3.up * localScale;
+            //SuesMandiblesConsumedPrefab.transform.localScale = Vector3.up * localScale;
+            //SilverThreadConsumedPrefab.transform.localScale = Vector3.up * localScale;
+            //UniversalSolventPrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Equipment * * * 
 
@@ -292,7 +298,7 @@ namespace UltitemsCyan
             YieldSignStopPrefab = mainBundle.LoadAsset<GameObject>("YieldSignStop.prefab");
 
             //JellyJailPrefab.transform.localScale = Vector3.up * localScale;
-            IceCubesPrefab.transform.localScale = Vector3.up * localScale;
+            //IceCubesPrefab.transform.localScale = Vector3.up * localScale;
             //PetRockPrefab.transform.localScale = Vector3.up * localScale;
             //TrebuchetPrefab.transform.localScale = Vector3.up * localScale;
 
@@ -306,20 +312,21 @@ namespace UltitemsCyan
             PotOfRegolithPrefab = mainBundle.LoadAsset<GameObject>("PotOfRegolith.prefab");
             UniversalSolutePrefab = mainBundle.LoadAsset<GameObject>("UniversalSolute.prefab");
 
-            MacroseismographPrefab.transform.localScale = Vector3.up * localScale;
-            MacroseismographConsumedPrefab.transform.localScale = Vector3.up * localScale;
-            PotOfRegolithPrefab.transform.localScale = Vector3.up * localScale;
-            UniversalSolutePrefab.transform.localScale = Vector3.up * localScale;
+            //MacroseismographPrefab.transform.localScale = Vector3.up * localScale;
+            //MacroseismographConsumedPrefab.transform.localScale = Vector3.up * localScale;
+            //PotOfRegolithPrefab.transform.localScale = Vector3.up * localScale;
+            //UniversalSolutePrefab.transform.localScale = Vector3.up * localScale;
 
             // * * * Buffs * * * 
             BirthdaySprite = mainBundle.LoadAsset<Sprite>("Birthday");
-            ChrysotopeFlySprite = mainBundle.LoadAsset<Sprite>("ChrysotopeFly");
+            CrysotopeFlySprite = mainBundle.LoadAsset<Sprite>("ChrysotopeFly");
             DownloadedSprite = mainBundle.LoadAsset<Sprite>("Downloaded");
             DreamSpeedSprite = mainBundle.LoadAsset<Sprite>("DreamSpeed");
             FrisbeeGlideSprite = mainBundle.LoadAsset<Sprite>("FrisbeeGlide");
             GrapeSprite = mainBundle.LoadAsset<Sprite>("Grape");
             OverclockedSprite = mainBundle.LoadAsset<Sprite>("Overclocked");
             RottingSprite = mainBundle.LoadAsset<Sprite>("Rotting");
+            SporkBleedSprite = mainBundle.LoadAsset<Sprite>("SporkBleed");
             TickCritSprite = mainBundle.LoadAsset<Sprite>("TickCrit");
             ZorseStarveSprite = mainBundle.LoadAsset<Sprite>("ZorseStarve");